+++ flexbot/config.go
@@ -5,7 +5,7 @@ import (
 	
 	"github.com/igor-feoktistov/terraform-provider-flexbot/pkg/rancher"
 	"github.com/igor-feoktistov/terraform-provider-flexbot/pkg/config"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	rancherManagementClient "github.com/rancher/rancher/pkg/client/generated/management/v3"
 )